:recycle: refactor windowModal place in project
