+++ src/entities/modal-window/index.ts
@@ -1,0 +1,3 @@
+export * from './lib/modal'
+
+export { default as ModalWindow } from './ui/ModalWindow.vue'
