+++ src/entities/modal-window/lib/modal.ts
