+++ src/entities/modal-window/ui/modal-window.scss
