+++ src/pages/password-entry/components/PasswordEntry.vue
@@ -1,11 +1,17 @@
 <script setup lang="ts">
 import { MasterForm } from '@/features/entry'
-import ModalWindow from '@/shared/ui/modal-window/modalWindow.vue'
-import { open } from '@/shared/lib/modal'
+import { ModalWindow } from '@/entities/modal-window/index'
+import { open, modalShow } from '@/entities/modal-window/index'
 </script>
 
 <template>
   <main :inert="open">
+    <button
+      aria-label="открыть модальное окно для ввода мастерпароля"
+      @click="e => modalShow(e, MasterForm, 'Register')"
+    >
+      открыть
+    </button>
     <h1>Password entry</h1>
     <MasterForm />
   </main>
