+++ src/pages/work-screen/components/WorkScreen.vue
@@ -3,8 +3,8 @@ import { ConfigurationPanel } from '@/widgets/configuration-panel'
 import { ServicesList } from '@/widgets/services-list'
 import { ShortHash } from '@/widgets/short-hash'
 import { exit } from '@/features/entry'
-import ModalWindow from '@/shared/ui/modal-window/modalWindow.vue'
-import { open } from '@/shared/lib/modal'
+import { ModalWindow } from '@/entities/modal-window/index'
+import { open } from '@/entities/modal-window/index'
 </script>
 
 <template>
